feat(api): lazy role-guarded agentWs import in the command relay — heavy agent-I/O workers move to the worker container (#4141)
The static value-import of routes/agentWs.ts in services/agentCommandRelay.ts
was the single edge pinning monitor/snmp/backup/discovery (and networkBaseline
via discovery) to socket-owner placement, keeping the 3.5d split's heaviest
workloads on the api container. The import is now a module-cached dynamic
import inside the two already-role-guarded branches; the closure contract test
gains a role-safe-boundary rule for the facade plus self-invalidating guards
(no static agentWs value-import; both call sites role-guarded). discovery and
networkBaseline had a second chain (isCronDue via automationRuntime →
scriptDispatch → agentWs) — the pure cron evaluator is extracted to
services/cronDue.ts with automationRuntime re-exporting.

Placement: 81 global / 24 socket-owner (was 76/29). Verified live: worker-role
boot against real PG+Redis reaches ready with 81 workers, relay integration
suite 9/9, SIGTERM walks the 7 phases to exit 0.
