+++ docs/superpowers/plans/ai-mcp/2026-08-27-ai-agents-wave3.5d-b-role-split.md
@@ -164,7 +164,7 @@ export function registerAiAgentEnqueuer(): void {
 - [x] **Step 2: Run it — let it FAIL on every misclassified entry, then flip those entries to `socket-owner` in the registry until honest-green.** Record the final classification (counts + the socket-owner name list) in THIS plan doc under "Final placement" below, and in the PR body. Sanity-expect: monitor/snmp/backup/discovery/networkBaseline ALL socket-owner, not just monitorWorker — every one of their closures reaches `routes/agentWs.ts` via `services/agentCommandRelay.ts:25`'s `import { isAgentConnected, sendCommandToAgent } from '../routes/agentWs'` — a VALUE import, not type-only, so the closure tool correctly counts it even though the calls it guards are runtime-safe (`breezeRole() !== 'worker'`-gated; Task 3's extraction only removed `routes/monitors.ts` from monitorWorker's closure, not this relay path). Consequence: these heavy agent-I/O workers stay in the `api` container THIS wave — `worker` never runs them. The one-line lazy-import flip (moving `agentCommandRelay.ts`'s `routes/agentWs` import behind an `await import(...)` inside its `breezeRole() !== 'worker'`-guarded branches) is the immediate follow-up that would let all five go `global`. automation/quickSupport/softwareDeployment/commandQueue-reachers socket-owner.
 - [x] **Step 3: Commit:** `test(api): worker entrypoint import-closure contract + final placements (#4086)`
 
-**Final placement (filled 2026-08-27, Task 5):** global = 76, socket-owner = 29 (out of 105 total). The closure tool confirmed every one of Task 1's initial-pass guesses except one: `policyEvaluationWorker` was mis-flagged `global` and is flipped to `socket-owner` here — its runtime closure reaches `routes/agentWs.ts` via a dynamic import one hop out (`jobs/policyEvaluationWorker.ts` → `services/policyEvaluationService.ts` → `await import('../jobs/automationWorker')` for `enqueueAutomationRun` → `services/automationRuntime.ts` → `services/scriptDispatch.ts` → `routes/agentWs.ts`). All other sanity-expected classifications held: monitor/snmp/backup/discovery/networkBaseline are ALL socket-owner (`services/agentCommandRelay.ts:25` value-imports `isAgentConnected`/`sendCommandToAgent` from `routes/agentWs.ts` — runtime-safe behind a `breezeRole() !== 'worker'` guard, but still a static edge the mechanical closure walk correctly counts); automation/quickSupport/softwareDeployment/commandQueue-reachers socket-owner. Consequence: this whole family of heavy agent-I/O workers stays on the `api`/`all` container this wave, not `worker` — migrating `agentCommandRelay.ts`'s `routes/agentWs` import to a lazy `await import(...)` inside its guarded branches (filed follow-up, see the "Do not relitigate" line above) is the one-line change that would let all five move to `global` in a later wave.
+**Final placement (filled 2026-08-27, Task 5):** global = 81, socket-owner = 24 (out of 105 total) — amended by #4141: services/agentCommandRelay.ts now lazy-imports routes/agentWs inside its role-guarded branches, and the closure contract test treats the facade as a role-safe boundary module (with self-invalidating guards: no static agentWs value-import, both call sites role-guarded), so discovery/networkBaseline/snmp/monitor/backup all flipped to global (networkBaseline+discovery additionally needed isCronDue extracted to services/cronDue.ts — their second agentWs chain ran through automationRuntime→scriptDispatch). The closure tool confirmed every one of Task 1's initial-pass guesses except one: `policyEvaluationWorker` was mis-flagged `global` and is flipped to `socket-owner` here — its runtime closure reaches `routes/agentWs.ts` via a dynamic import one hop out (`jobs/policyEvaluationWorker.ts` → `services/policyEvaluationService.ts` → `await import('../jobs/automationWorker')` for `enqueueAutomationRun` → `services/automationRuntime.ts` → `services/scriptDispatch.ts` → `routes/agentWs.ts`). All other sanity-expected classifications held: monitor/snmp/backup/discovery/networkBaseline are ALL socket-owner (`services/agentCommandRelay.ts:25` value-imports `isAgentConnected`/`sendCommandToAgent` from `routes/agentWs.ts` — runtime-safe behind a `breezeRole() !== 'worker'` guard, but still a static edge the mechanical closure walk correctly counts); automation/quickSupport/softwareDeployment/commandQueue-reachers socket-owner. Consequence: this whole family of heavy agent-I/O workers stays on the `api`/`all` container this wave, not `worker` — migrating `agentCommandRelay.ts`'s `routes/agentWs` import to a lazy `await import(...)` inside its guarded branches (filed follow-up, see the "Do not relitigate" line above) is the one-line change that would let all five move to `global` in a later wave.
 
 socket-owner (28): fleetRemediationDispatchWorker, policyEvaluationWorker, softwareComplianceWorker, softwareRemediationWorker, aiAgentRunner, auditBaselineJobs, cisJobs, automationWorker, backupVerificationJobs, eventLogRetention, quickSupportReaper, desktopSessionFinalization, desktopSessionOrphanRecovery, discoveryWorker, networkBaselineWorker, snmpWorker, monitorWorker, patchJobWorker, patchSchedulerWorker, maintenanceRebootWorker, backupWorker, sensitiveDataWorker, peripheralJobs, drExecutionWorker, staleCommandReaper, softwareDeploymentScheduler, offboardingDrainReaper, intentReleaseWorker.
